requirements.txt
pytesseract==0.3.13
pdf2image==1.17.0
transformers==4.46.3
torch==2.5.1
accelerate==1.1.1
Pillow==11.0.0
requests==2.32.3